requirements.txt
coverage
gevent
mock
nose
pika==1.1.0
pyyaml
aio-pika==6.1.2